git-p4: add format-patch subcommand
It takes a list of P4 changelists and generates a patch for
each one, using "p4 describe".

e.g.

  $ git p4 format-patch -o patches 1 2 3 4

This is especially useful for applying shelved changelists
to your git-p4 tree; the existing "git p4" subcommands do
not handle these.

That's because they "p4 print" the contents of each file at
a given revision, and then use git-fastimport to generate the
deltas. But in the case of a shelved changelist, there is no
easy way to find out what the previous file state was - Perforce
does not have the concept of a single repo-wide revision.

Signed-off-by: Luke Diamand <luke@diamand.org>