git-p4: add P4FileInfo helper class
This introduces a small helper class, P4FileInfo, for dealing
with the dictionaries of file info results produced by P4
commands such as "describe".

The dictionary is really a list of dictionaries, but the keys
are concatenated with the index. For example:

    depotPath0 : 'foo'
    type0 : 'text'
    rev0 : 1
    depotPath1 : 'bar'
    type1 : 'text'
    rev1 : 2

An instance of this class is one slice of that list. It is still
a dictionary, so is a drop-in replacement.