git-p4: better error reporting if perforce is inaccessible
It might not be possible to contact the Perforce server: the
usual reasons are that you've been logged out, or P4PORT is
not set, or the p4 binary is not in your path.

When this happens, currently git-p4 crashes with an obscure
error message.

Instead, check if we can talk to Perforce before running
the first command, and if not, try to report a useful error.

Signed-off-by: Luke Diamand <luke@diamand.org>