git-p4: use a dictionary which can automatically convert strings to bytes
git-p4 receives marshalled data which has key names in roman-1, and data
as some kind of binary format, which may or may not need decoding.

The current codebase assumes python2 behavior where this data can be
simply used as-is.

In python3, the keys are now byte strings rather than strings.

To handle this in a way which will work in both python3 and python2,
this change adds a "PolyStringDict" which automatically converts keys
to/from string/bytes as needed. This is loosely inspired by the
PolyString approach used in reposurgeon.

Signed-off-by: Luke Diamand <luke@diamand.org>