git-p4: fully support unshelving changelists
The previous unshelve support constructed the new git commit on
top of the current HEAD revision. If that revision didn't match
the tree at the point where the P4 changelist was shelved, it would
complain and give up, since otherwise the user would end up with an
unshelved commit containing bits of other intervening commits.

e.g. given fileA:
      the
      quick
      brown
      fox

  change1: s/the/The/   <- p4 shelve this change
  change2: s/fox/Fox/   <- p4 submit this change
  git p4 unshelve 1     <- FAIL

This change teaches the P4Unshelve class to always create a parent
commit which matches the P4 tree at the point prior to the P4
shelve being created (which is handily reported in the p4 description
for a shelved changelist).

That then means git-p4 can always create a git commit matching the
P4 shelve that was originally created.

Additionally extends the tests to check for p4 move operations.

Signed-off-by: Luke Diamand <luke@diamand.org>