git-p4: add functions to encode/decode in Python3
In Python2, everything is a string, but in Python3, strings
are unicode and are not the same as arrays of bytes, which require
an encoding to make sense.

Git uses an encoding of utf-8. P4 will use utf-8 when talking to a
unicode server, but can also use other encodings in some circumstances.

https://community.perforce.com/s/article/3104

This change just adds some helper functions to encode/decode to/from
strings and bytes for git and python.